LayoutTests/fast/dom/xmlserializer-ampersand-in-style-expected.txt

Tests XML serialization of ampersand in a style element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test case: /* & < > "  */
PASS result is "<style xmlns=\"http://www.w3.org/1999/xhtml\">/* &amp; &lt; &gt; \"  */</style>"
PASS successfullyParsed is true

TEST COMPLETE
